restructured notes:
