+++ lib/Screens/Note/NoteDetail.dart
@@ -4,6 +4,7 @@ import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 import 'package:untitled3/Observables/NoteObservable.dart';
 import 'package:untitled3/Services/NoteService.dart';
+import 'package:untitled3/Utility/Constant.dart';
 import 'package:untitled3/generated/i18n.dart';
 
 class NoteDetails extends StatefulWidget {
@@ -169,7 +170,7 @@ class _NoteDetailssState extends State<NoteDetails> {
       ),
       onPressed: () {
         noteObserver.deleteNote(noteObserver.currNoteForDetails);
-        noteObserver.changeScreen(I18n.of(context)!.notesScreenName);
+        noteObserver.changeScreen(NOTE_SCREENS.NOTE);
         Navigator.of(context).pop();
       },
     );
