+++ lib/Screens/Note/SaveNote.dart
@@ -3,6 +3,7 @@ import 'package:fluttertoast/fluttertoast.dart';
 import 'package:provider/provider.dart';
 import 'package:untitled3/Model/Note.dart';
 import 'package:untitled3/Services/NoteService.dart';
+import 'package:untitled3/Utility/Constant.dart';
 import 'package:untitled3/generated/i18n.dart';
 import '../../Observables/NoteObservable.dart';
 
@@ -83,7 +84,7 @@ class _SaveNoteState extends State<SaveNote> {
                     note.text = textController.text;
                     noteObserver.addNote(note);
                     showToast();
-                    noteObserver.changeScreen(I18n.of(context)!.notesScreenName);
+                    noteObserver.changeScreen(NOTE_SCREENS.NOTE);
                   }
                 },
                 child: Text('Save'),
